requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
Django==2.1.15
djangorestframework==3.11.1
future==0.18.2
geocoder==1.38.1
gunicorn==20.0.4
idna==2.10
joblib==0.16.0
nltk==3.5
numpy==1.19.1
oauthlib==3.1.0
PySocks==1.7.1
pytz==2020.1
ratelim==0.1.6
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
sqlparse==0.3.1
textblob==0.15.3
threadpoolctl==2.1.0
tqdm==4.48.2
tweepy==3.9.0
urllib3==1.25.10
whitenoise==5.2.0
